Extract depositing logic into separate protocol
For now, this allows us to more easily mock the depositing.
